(Huge CL) Update Cobalt to use m70 Chromium repos
This is part of the project to update Cobalt's Chromium repos. Now
everything can compile and Cobalt is usable. A few tests are still
failing.

There is no need to go through everything, but please review these files:
base/wrap_main_starboard.h
browser/trace_manager.h(deleted)
debug/remote/debug_web_server.*
dom/html_image_element.*
h5vcc/dial/*
script/callback_function.h
math/r_tree_base.h
math/r_tree_unittest.cc
media/fetcher_buffered_data_source.*
network/cookie_jar_impl.*
network/network_delegate.*
netowrk/network_module.*
network/net_log_logger.*
network/url_request_context.*
network/persistent_cookie_store.*
network/cobalt_net_log.*
trace_event/scoped_event_parser_trace.*
script/union_type.h
loader/cobalt_url_response_string_writer.*
loader/net_fetcher.*
websocket/*(deleted and modified many files)
xhr/xml_http_request.cc

b/27798606

+++ cobalt/css_parser/position_parse_structures.cc
@@ -135,7 +135,7 @@ bool PositionParseStructure::IsPositionValidOnSizeTwo() {
   // Check for the special case that a pair of keywords can be reordered while
   // a combination of keyword and length or percentage cannot. eg. 'top center'
   // is valid while 'top 50%' is not.
-  DCHECK_EQ(2, position_builder_.size());
+  DCHECK_EQ(size_t(2), position_builder_.size());
 
   Direction first_direction = ConvertToDirection(position_builder_[0]);
   Direction second_direction = ConvertToDirection(position_builder_[1]);
